@@ -22,26 +22,35 @@
   "homepage": "https://github.com/thedreadpirate/overseer",
   "dependencies": {
     "cal-heatmap": "3.5.4",
+    "cldr-data": "latest",
     "d3": "3.5.6",
     "flux": "2.1.1",
-    "griddle-react": "0.2.14",
-    "jest-cli": "0.5.4",
+    "globalize": "1.x",
+    "griddle-react": "0.8.x",
+    "history": "3.x",
     "moment": "2.10.6",
     "object-assign": "3.0.0",
-    "react": "0.13.3",
-    "react-notification-system": "0.1.14",
-    "react-router": "0.13.3",
-    "react-skylight": "0.2.0",
-    "react-widgets": "2.8.0",
+    "prop-types": "15.x",
+    "react": "16.x",
+    "react-dom": "16.x",
+    "react-notification-system": "0.2.x",
+    "react-router": "3.x",
+    "react-skylight": "0.5.1",
+    "react-widgets": "4.x",
+    "react-widgets-globalize": "latest",
     "underscore": "1.8.3"
   },
   "devDependencies": {
-    "babelify": "6.3.0",
-    "browserify": "11.0.1",
+    "babel-core": "6.x",
+    "babelify": "8.x",
+    "babel-preset-stage-1": "latest",
+    "babel-preset-es2015": "6.x",
+    "babel-preset-react": "6.x",
+    "browserify": "16.x",
     "concurrently": "2.0.0",
-    "reactify": "1.1.1",
-    "uglifyify": "3.0.1",
-    "watchify": "3.4.0",
-    "npm-shrinkwrap": "6.0.1"
+    "jest-cli": "0.5.4",
+    "reactify": "1.x",
+    "uglifyify": "4.0.x",
+    "watchify": "3.x"
   }
 }
