@@ -1,28 +1,35 @@
 var React = require('react'),
-    Skylight = require('react-skylight');
+    Skylight = require('react-skylight').default;
 
-module.exports = React.createClass({
-    show: function () {
+module.exports = class extends React.Component {
+    static displayName = 'Modal';
+
+    show = () => {
         this.refs.modal.show();
         $(document.body).off('keydown');
         $(document.body).on('keydown', this.handleKeyDown);
-    },
-    hide: function () {
+    };
+
+    hide = () => {
         this.refs.modal.hide();
-    },
-    handleKeyDown: function (keypress) {
+    };
+
+    handleKeyDown = (keypress) => {
         if (keypress.keyCode == 27 /*esc*/) {
             this.hide();
             this.unbindEsc();
         }
-    },
-    unbindEsc: function(){
+    };
+
+    unbindEsc = () => {
         $(document.body).off('keydown');
-    },
-    componentWillUnMount: function () {
+    };
+
+    componentWillUnMount = () => {
         this.unbindEsc();
-    },
-    render: function () {
+    };
+
+    render() {
         return <Skylight ref="modal" dialogStyles={{height: '450px', backgroundColor: '', boxShadow: ''}}>
             <div className="inner-large-content panel panel-primary" style={{height: '100%'}}>
                 <div className="panel-heading">{this.props.title}</div>
@@ -32,5 +39,5 @@ module.exports = React.createClass({
             </div>
         </Skylight>;
     }
-});
+};
 
