@@ -5,41 +5,40 @@ var React = require('react'),
     SuperWrapper = require('./superwrapper.jsx'),
     AdminWrapper = require('./adminwrapper.jsx');
 
-module.exports = React.createClass({
+module.exports = class Nav extends React.Component {
+    state = {selectedSchool: userStore.getSelectedSchool()};
 
-    getInitialState: function () {
-        return {selectedSchool: userStore.getSelectedSchool()};
-    },
-
-    componentDidMount: function () {
+    componentDidMount() {
         userStore.addChangeListener(this._onChange);
-    },
-    componentWillUnmount: function () {
-        userStore.removeChangeListener(this._onChange);
-    },
+    }
 
-    _onChange: function () {
+    componentWillUnmount() {
+        userStore.removeChangeListener(this._onChange);
+    }
+
+    _onChange = () => {
         this.setState({selectedSchool: userStore.getSelectedSchool()});
-    },
-    render: function () {
+    };
+
+    render() {
         return <nav className="navbar" role="navigation">
             <div className="container">
                 <div className="navbar-header">
-                    <Link to="students" id="home" className="navbar-brand">{this.state.selectedSchool?this.state.selectedSchool.name:""} Custodia &mdash; Home</Link>
+                    <Link to="/students" id="home" className="navbar-brand">{this.state.selectedSchool?this.state.selectedSchool.name:""} Custodia &mdash; Home</Link>
                 </div>
                 <div id="navbar">
                     <ul className="nav navbar-nav">
                         <AdminWrapper>
-                            <li><Link id="totals-link" to="reports">Reports</Link></li>
+                            <li><Link id="totals-link" to="/reports">Reports</Link></li>
                         </AdminWrapper>
                         <AdminWrapper>
-                            <li><Link id="students-link" to="studentAdmin">Students</Link></li>
+                            <li><Link id="students-link" to="/studentAdmin">Students</Link></li>
                         </AdminWrapper>
                         <AdminWrapper>
-                            <li><Link id="class-link" to="classes">Classes</Link> </li>
+                            <li><Link id="class-link" to="/classes">Classes</Link> </li>
                         </AdminWrapper>
                         <SuperWrapper>
-                            <li><Link id="class-link" to="admin">Site Admin</Link> </li>
+                            <li><Link id="class-link" to="/administration">Site Admin</Link> </li>
                         </SuperWrapper>
                     </ul>
                     <ul className="nav navbar-nav">
@@ -49,4 +48,4 @@ module.exports = React.createClass({
             </div>
         </nav>;
     }
-});
+};
