@@ -1,35 +1,44 @@
 var React = require('react'),
+    PropTypes = require('prop-types'),
     Router = require('react-router'),
     Link = Router.Link,
     AdminItem = require('./adminwrapper.jsx'),
     actionCreator = require('./studentactioncreator'),
+    myhistory = require('./myhistory.js'),
     SwipeHelpers = require('./swipeHelpers.jsx'),
     studentStore = require('./StudentStore');
 
-module.exports = React.createClass({
-    contextTypes: {
-        router: React.PropTypes.func.isRequired
-    },
-    getInitialState: function () {
-        return {students: studentStore.getStudents(true),
-                today: null};
-    },
-    componentDidMount: function () {
+module.exports = class extends React.Component {
+    static displayName = "StudentTable";
+
+    static contextTypes = {
+        router: PropTypes.object.isRequired
+    };
+
+    state = {students: studentStore.getStudents(true),
+            today: null};
+
+    componentDidMount() {
         studentStore.addChangeListener(this._onChange);
-    },
-    componentWillUnmount: function () {
+    }
+
+    componentWillUnmount() {
         studentStore.removeChangeListener(this._onChange);
-    },
-    signIn: function (student) {
+    }
+
+    signIn = (student) => {
         this.refs.missingSwipeCollector.validateSignDirection(student, 'in');
-    },
-    signOut: function (student) {
+    };
+
+    signOut = (student) => {
         this.refs.missingSwipeCollector.validateSignDirection(student, 'out');
-    },
-    isSigningIn: function(student) {
-        return !student.last_swipe_date || student.last_swipe_type === 'out' || !student.last_swipe_date.startsWith(studentStore.getToday());
-    },
-    getSwipeButton:  function(student, way){
+    };
+
+    isSigningIn = (student) => {
+        return student.last_swipe_type === 'out' || !student.in_today;
+    };
+
+    getSwipeButton = (student, way) => {
         var buttonIcon = 'fa-arrow-right';
         if (way === 'out') {
             buttonIcon = 'fa-arrow-left';
@@ -52,14 +61,14 @@ module.exports = React.createClass({
                 <i className={iclassName}>&nbsp;</i>
              </button>;
         }
-    },
-    getStudent: function(student, way){
+    };
+
+    getStudent = (student, way) => {
         var link = <span className="glyphicon glyphicon-calendar"></span>;
         var button = this.getSwipeButton(student, way);
         var calendar_button_class = "btn btn-default calendar-button";
-        var self = this;
         var calendar_button = <div onClick={function() {
-            self.context.router.transitionTo('student', {studentId: student._id})}}
+            myhistory.push('/students/' + student._id); }}
             className={calendar_button_class}>{link}</div>;
 
         if (way !== 'out') {
@@ -73,8 +82,9 @@ module.exports = React.createClass({
                 {calendar_button}
             </div>;
         }
-    },
-    render: function () {
+    };
+
+    render() {
         var absentCol = [],
             notYetInCol = [],
             inCol = [],
@@ -128,9 +138,10 @@ module.exports = React.createClass({
                 </div>
             </div>
         </div>;
-    },
-    _onChange: function () {
+    }
+
+    _onChange = () => {
         this.setState({students: studentStore.getStudents(),
                        today: studentStore.getToday()});
-    }
-});
+    };
+};
