@@ -1,12 +1,14 @@
 var React = require('react'),
+  PropTypes = require('prop-types'),
   AdminItem = require('./adminwrapper.jsx'),
   actionCreator = require('./studentactioncreator');
 
-var exports = React.createClass({
-    contextTypes: {
-        router: React.PropTypes.func
-    },
-    getCurrentDay: function (student, dayString) {
+class SwipesListing extends React.Component {
+    static contextTypes = {
+        router: PropTypes.object
+    };
+
+    getCurrentDay = (student, dayString) => {
         if(student) {
             var day = student.days.find(function (day) {
                 return day.day === dayString;
@@ -15,21 +17,22 @@ var exports = React.createClass({
         }else{
             return {};
         }
-    },
-    componentWillReceiveProps: function(newProps){
+    };
+
+    componentWillReceiveProps(newProps) {
       this.setState({day: this.getCurrentDay(newProps.student, newProps.day)});
-    },
-    getInitialState: function(){
-        return {day: this.getCurrentDay(this.props.student, this.props.day)};
-    },
-    deleteSwipe: function (swipe) {
+    }
+
+    deleteSwipe = (swipe) => {
         actionCreator.deleteSwipe(swipe, this.props.student);
-    },
-    swipesAreEmpty: function(swipes){
+    };
+
+    swipesAreEmpty = (swipes) => {
         return swipes.length === 0 ||
             (swipes.length === 1 && swipes[0].in_time === null && swipes[0].out_time === null);
-    },
-    getSwipesForDay: function () {
+    };
+
+    getSwipesForDay = () => {
         var swipeRows = [];
         if (this.state.day && !this.swipesAreEmpty(this.state.day.swipes)) {
             this.state.day.swipes.map(function (swipe) {
@@ -47,14 +50,19 @@ var exports = React.createClass({
             return <tr><td colSpan="2">No swipes available.</td></tr>;
         }
         return swipeRows;
-    },
-    excuse: function (swipe) {
+    };
+
+    excuse = (swipe) => {
         actionCreator.excuse(this.props.student._id, this.props.day);
-    },
-    override: function () {
+    };
+
+    override = () => {
         actionCreator.override(this.props.student._id, this.props.day);
-    },
-    render: function () {
+    };
+
+    state = {day: this.getCurrentDay(this.props.student, this.props.day)};
+
+    render() {
         this.state.day.round_mins = parseFloat(this.state.day.total_mins).toFixed(0);
         var showOverrideExcuseButtons = this.state.day && !this.state.day.override
                                  && !this.state.day.excused && !this.state.day.valid;
@@ -79,22 +87,16 @@ var exports = React.createClass({
             {showOverrideExcuseButtons ?
               <AdminItem>
                 <div className="action-buttons">
-                  <div className="pull-left">
-                    <button type="button" id="override" onClick={this.override} className="btn btn-sm btn-info">
-                        Override
-                    </button>
-                  </div>
-                  <div className="pull-right">
-                    <button type="button" onClick={this.excuse} className="btn btn-sm btn-info">Excuse</button>
-                  </div>
+                  <button type="button" id="override" onClick={this.override} className="btn btn-sm btn-info">
+                      Override
+                  </button>
+                  <button type="button" onClick={this.excuse} className="btn btn-sm btn-info">Excuse</button>
                 </div>
               </AdminItem>
               : ''
             }
             </span>;
-    },
+    }
+}
 
-
-});
-
-module.exports = exports;
+module.exports = SwipesListing;
